Add Jacobian to quadrature module functions
Previously the Jacobian of the spatial transformation had to be
incorporated to the integration manually, now it's an argument of the
integration and averaging quadrature functions for a cell.

All calls to quadrature functions were updated along with relevant
tests.

Also a possible bug in the projection limiter might have been fixed.
From the comments it seems like the reference implementation of the
limiter in the GFR code used the jacobian in the integration process for
the marker. This was not done in FREI previosly but is now.