Feat: Change stdout convergence tracking metric
Swap the old Ratio of L2-Norms of the solution delta in the current iteration
divided by the delta in the first iteration for just the L2-Norm of the delta
in the current iteration. This makes it easier to compare convergence of
different problems and now that the solver is non-dimensional we should have
all solution deltas and residues close to the same order of magnitude.