Fix: Interpolate Sol from SPs to FPs at output
Previously we just used the latest value assigned to the solution at the
FP, but this value is out of date since the solution is updated at the
end of the RK step. It becomes a deal breaker when we try to recover the
last solution before the solver diverges which is a planed feature.